+++ src/xsalsa20.c
@@ -167,13 +167,13 @@ void conch_xsalsa20_block(struct xsalsa20_ctx *ctx, int32_t n)
 void conch_xsalsa20_crypto(struct xsalsa20_ctx *ctx, uint8_t *buf,
 		size_t len)
 {
-	uint32_t n = XSALSA20_BLOCKSIZE;
+	size_t n = XSALSA20_BLOCKSIZE;
 	while (len) {
 		if (len < n)
 			n = len;
 
 		conch_xsalsa20_block(ctx, XSALSA20_ROUNDS);
-		for (uint32_t i = 0; i < n; i++)
+		for (size_t i = 0; i < n; i++)
 			buf[i] ^= XSALSA20_KEYSTREAM(ctx, i);
 
 		buf += n;
